Split color and canvas
